backend-python/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pandas
yfinance
apscheduler
openai
python-dotenv
numpy
python-multipart
scipy
psycopg2-binary
